MAKING LOCALHOST/CARS ( AKA INDEX )
Added a new shared folders for html that we wanna render in several pages.
In this case i added footer and navbar to the shared folder and called render
on application html.

Navbar, footer and cards all have specific scss folders. Right now the styling
is not finished but the main structure is halfway thro. Still gotta fix the geomap
div so tomorrow after lesson we will be able to instantly add geomap to the div.