PG SEARCH FUNCTIONALITY
Warning: I killed(commented out) the Landing page just for better
code approach while testing.
Right now the landing page is searching by city but we need to test
after having all the seeds if the markers aren't fucked up. IN case
they aren't working as they should we should just duplicate the marker
code on car controller and put it inside the if and the else.